refactor(stellar): split StellarService (588 lines) into focused services
Extract cohesive responsibilities from the monolithic StellarService into
separate focused services following the SorobanQuestReaderService pattern:

- StellarSubmissionService: Soroban contract submission (approveSubmission,
  signAndSubmit, _signAndSubmitContract)
- StellarPaymentService: native XLM/asset payment transfers (sendPayment)
- StellarEventIngestionService: cron-based Soroban event ingestion with
  deduplication (ingestContractEvents)

StellarService is now a slim infrastructure provider exposing getHorizon(),
getRpc(), and getNetworkPassphrase() to the focused services.

Updated DI wiring in stellar.module.ts to register and export all new
services. Consumers updated: SubmissionsService → StellarSubmissionService,
PayoutProcessor → StellarPaymentService.

Added spec coverage for all new services. Updated all affected test files.

Closes #1912
