+++ BackEnd/src/modules/stellar/stellar.module.ts
@@ -2,12 +2,27 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { StellarService } from './stellar.service';
+import { StellarSubmissionService } from './stellar-submission.service';
+import { StellarPaymentService } from './stellar-payment.service';
+import { StellarEventIngestionService } from './stellar-event-ingestion.service';
 import { SorobanQuestReaderService } from './soroban-quest-reader.service';
 import { EventStore } from '../../events/entities/event-store.entity';
 
 @Module({
   imports: [ConfigModule, TypeOrmModule.forFeature([EventStore])],
-  providers: [StellarService, SorobanQuestReaderService],
-  exports: [StellarService, SorobanQuestReaderService],
+  providers: [
+    StellarService,
+    StellarSubmissionService,
+    StellarPaymentService,
+    StellarEventIngestionService,
+    SorobanQuestReaderService,
+  ],
+  exports: [
+    StellarService,
+    StellarSubmissionService,
+    StellarPaymentService,
+    StellarEventIngestionService,
+    SorobanQuestReaderService,
+  ],
 })
 export class StellarModule {}
