+++ BackEnd/src/modules/submissions/submissions.service.ts
@@ -27,7 +27,7 @@ import { Submission, SubmissionStatus } from './entities/submission.entity';
 import { ApproveSubmissionDto } from './dto/approve-submission.dto';
 import { RejectSubmissionDto } from './dto/reject-submission.dto';
 import { CreateSubmissionDto } from './dto/create-submission.dto';
-import { StellarService } from '../stellar/stellar.service';
+import { StellarSubmissionService } from '../stellar/stellar-submission.service';
 import { NotificationsService } from '../notifications/notifications.service';
 import { Quest } from '../quests/entities/quest.entity';
 import { User } from '../users/entities/user.entity';
@@ -60,7 +60,7 @@ export class SubmissionsService {
     private usersRepository: Repository<User>,
     @InjectRepository(Quest)
     private questsRepository: Repository<Quest>,
-    private stellarService: StellarService,
+    private stellarSubmissionService: StellarSubmissionService,
     private notificationsService: NotificationsService,
     private eventEmitter: EventEmitter2,
     private metricsService: MetricsService,
@@ -301,7 +301,7 @@ export class SubmissionsService {
 
     let onChainTxHash: string | undefined;
     try {
-      const onChainResult = await this.stellarService.approveSubmission(
+      const onChainResult = await this.stellarSubmissionService.approveSubmission(
         quest.contractTaskId,
         user.stellarAddress,
         verifier.stellarAddress,
