+++ BackEnd/test/integration/full-application.integration-spec.ts
@@ -22,7 +22,8 @@ import { UsersService } from '#src/modules/users/users.service';
 import { QuestsService } from '#src/modules/quests/quests.service';
 import { SubmissionsService } from '#src/modules/submissions/submissions.service';
 import { PayoutsService } from '#src/modules/payouts/payouts.service';
-import { StellarService } from '#src/modules/stellar/stellar.service';
+import { StellarSubmissionService } from '#src/modules/stellar/stellar-submission.service';
+import { StellarPaymentService } from '#src/modules/stellar/stellar-payment.service';
 
 // Import entities
 import { User } from '#src/modules/users/entities/user.entity';
@@ -90,16 +91,17 @@ describe('Full Application Integration', () => {
           .mockResolvedValue({ stellarAddress: 'test', sub: 'test' }),
         decode: jest.fn(),
       })
-      .overrideProvider(StellarService)
+      .overrideProvider(StellarSubmissionService)
       .useValue({
         approveSubmission: jest
           .fn()
           .mockResolvedValue({ transactionHash: 'tx-hash-mock' }),
+      })
+      .overrideProvider(StellarPaymentService)
+      .useValue({
         sendPayment: jest
           .fn()
           .mockResolvedValue({ transactionHash: 'tx-hash-mock' }),
-        getContractId: jest.fn().mockReturnValue('mock-contract-id'),
-        getServer: jest.fn(),
       })
       .compile();
